backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
django-cors-headers>=4.4,<5
gunicorn>=22.0,<23
